+++ wakamiti/wakamiti-api/src/main/java/es/iti/wakamiti/api/WakamitiAPI.java
@@ -37,4 +37,5 @@ public interface WakamitiAPI {
     String version();
     Path workingDir(Configuration configuration);
 
+    Configuration defaultConfiguration();
 }