+++ wakamiti/wakamiti-api/src/main/java/es/iti/wakamiti/api/WakamitiStepRunContext.java
@@ -43,8 +43,7 @@ public class WakamitiStepRunContext {
     private final Backend backend;
     private final Locale stepLocale;
     private final Locale dataLocale;
-    private final ResourceLoader resourceLoader;
-    private final Path workingDir;
+
 
 
     public WakamitiStepRunContext(
@@ -57,8 +56,6 @@ public class WakamitiStepRunContext {
         this.backend = backend;
         this.stepLocale = stepLocale;
         this.dataLocale = dataLocale;
-        this.workingDir = WakamitiAPI.instance().workingDir(configuration);
-        this.resourceLoader = WakamitiAPI.instance().resourceLoader(workingDir.toFile());
     }
 
 
@@ -86,27 +83,4 @@ public class WakamitiStepRunContext {
     }
 
 
-    public ResourceLoader resourceLoader() {
-        return resourceLoader;
-    }
-
-    public Path workingDir() {
-        return workingDir;
-    }
-
-
-    public File absolutePath(File file) {
-        if (file.isAbsolute()) {
-            return file;
-        }
-        return new File(workingDir.toFile(),file.toString());
-    }
-
-
-    public Path absolutePath(Path path) {
-        if (path.isAbsolute()) {
-            return path;
-        }
-        return workingDir.resolve(path);
-    }
 }